feat: implement rating repo logic

+++ internal/domain/ports/rating.ports.go
@@ -6,7 +6,7 @@ type RatingOutput interface {
 	Create(data entities.Rating) (entities.Rating, *entities.Exception)
 	Update(id string, data entities.Rating) (entities.Rating, *entities.Exception)
 	Delete(id string) *entities.Exception
-	List(listing entities.Listing) ([]entities.Rating, int64, *entities.Exception)
+	List(listing entities.Listing) ([]entities.Rating, int, *entities.Exception)
 	FindOne(id string) (entities.Rating, *entities.Exception)
 }
 
@@ -14,6 +14,6 @@ type RatingInput interface {
 	Create(data entities.Rating) (entities.Rating, *entities.Exception)
 	Update(id string, data entities.Rating) (entities.Rating, *entities.Exception)
 	Delete(id string) *entities.Exception
-	List(listing entities.Listing) ([]entities.Rating, int64, *entities.Exception)
+	List(listing entities.Listing) ([]entities.Rating, int, *entities.Exception)
 	FindOne(id string) (entities.Rating, *entities.Exception)
 }
